feat: extract ai handlers, kb stream
